Declaring constants to avoid dependency on newer API version
The constants for Play and Pause were added in API version 11 (the
first Honeycomb release). In order to avoid adding a dependency on this
newer, but less widely-used version, I'm manually declaring constants to
match those new values. If we ever to make an ICS-specific version, we
can remove this change.